// WorkSpace/projects/user/src/app/models/Project/iproject-details.ts
export interface IProjectDetails {
  id: number;
  nameAr: string;
  nameEn: string;
  descriptionAr: string;
  descriptionEn: string;
  addressAr: string;
  addressEn: string;
  image: string;
  siteEngineerNameAr: string;
  siteEngineerNameEn: string;
  projectProgressPercentage: number;
  companyId: number;
}